requirements.txt
alembic==1.13.2
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
marshmallow==3.21.3
passlib==1.7.4
psycopg2==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.31